@@ -1,4 +1,4 @@
-exports.getPosts = (req, res, next) => {
+export function getPosts(req, res, next) {
   res.status(200).json({
     posts: [
       {
@@ -13,9 +13,9 @@ exports.getPosts = (req, res, next) => {
       }
     ]
   });
-};
+}
 
-exports.createPost = (req, res, next) => {
+export function createPost(req, res, next) {
   const title = req.body.title;
   const content = req.body.content;
   // Create post in db
@@ -23,4 +23,4 @@ exports.createPost = (req, res, next) => {
     message: 'Post created successfully!',
     post: { id: new Date().toISOString(), title: title, content: content }
   });
-};
+}
